requirements.txt
nonebot2[fastapi]>=2.0.0
nonebot-adapter-onebot>=2.0.0
apscheduler>=3.10.0
feedparser>=6.0.0
aiohttp>=3.8.0
google-genai>=0.1.0
Pillow>=10.0.0
